requirements.txt
Django==1.8.12
mysqlclient==1.3.7
whitenoise==3.0
django-ckeditor==5.0.3
